+++ unstructured/partition/ocr.py
@@ -189,7 +189,6 @@ def supplement_page_layout_with_ocr(
     """
 
     ocr_agent = get_ocr_agent()
-    ocr_layout = None
     if ocr_mode == OCRMode.FULL_PAGE.value:
         ocr_layout = get_ocr_layout_from_image(
             image,
@@ -203,7 +202,8 @@ def supplement_page_layout_with_ocr(
     elif ocr_mode == OCRMode.INDIVIDUAL_BLOCKS.value:
         for element in page_layout.elements:
             if element.text == "":
-                padded_element = pad_element_bboxes(element, padding=12)
+                padding = env_config.IMAGE_CROP_PAD
+                padded_element = pad_element_bboxes(element, padding=padding)
                 cropped_image = image.crop(
                     (
                         padded_element.bbox.x1,
@@ -229,18 +229,12 @@ def supplement_page_layout_with_ocr(
     # Note(yuming): use the OCR data from entire page OCR for table extraction
     if infer_table_structure:
         table_agent = init_table_agent()
-        if ocr_layout is None:
-            # Note(yuming): ocr_layout is None for individual_blocks ocr_mode
-            ocr_layout = get_ocr_layout_from_image(
-                image,
-                ocr_languages=ocr_languages,
-                ocr_agent=ocr_agent,
-            )
         page_layout.elements[:] = supplement_element_with_table_extraction(
             elements=cast(List[LayoutElement], page_layout.elements),
-            ocr_layout=ocr_layout,
             image=image,
             table_agent=table_agent,
+            ocr_languages=ocr_languages,
+            ocr_agent=ocr_agent,
         )
 
     return page_layout
@@ -248,9 +242,10 @@ def supplement_page_layout_with_ocr(
 
 def supplement_element_with_table_extraction(
     elements: List[LayoutElement],
-    ocr_layout: List[TextRegion],
     image: PILImage,
     table_agent: "UnstructuredTableTransformerModel",
+    ocr_languages: str = "eng",
+    ocr_agent: str = OCR_AGENT_TESSERACT,
 ) -> List[LayoutElement]:
     """Supplement the existing layout with table extraction. Any Table elements
     that are extracted will have a metadata field "text_as_html" where
@@ -268,59 +263,38 @@ def supplement_element_with_table_extraction(
                     padded_element.bbox.y2,
                 ),
             )
-            table_tokens = get_table_tokens_per_element(
-                padded_element,
-                ocr_layout,
+            table_tokens = get_table_tokens(
+                image=cropped_image, ocr_languages=ocr_languages, ocr_agent=ocr_agent
             )
             element.text_as_html = table_agent.predict(cropped_image, ocr_tokens=table_tokens)
     return elements
 
 
-def get_table_tokens_per_element(
-    table_element: LayoutElement,
-    ocr_layout: List[TextRegion],
+def get_table_tokens(
+    image: PILImage,
+    ocr_languages: str = "eng",
+    ocr_agent: str = OCR_AGENT_TESSERACT,
 ) -> List[Dict]:
-    """
-    Extract and prepare table tokens within the specified table element
-    based on the OCR layout of an entire image.
+    """Get OCR tokens from either paddleocr or tesseract"""
 
-    Parameters:
-    - table_element (LayoutElement): The table element for which table tokens
-      should be extracted. It typically represents the bounding box of the table.
-    - ocr_layout (List[TextRegion]): A list of TextRegion objects representing
-      the OCR layout of the entire image.
-
-    Returns:
-    - List[Dict]: A list of dictionaries, each containing information about a table
-      token within the specified table element. Each dictionary includes the
-      following fields:
-        - 'bbox': A list of four coordinates [x1, y1, x2, y2]
-                    relative to the table element's bounding box.
-        - 'text': The text content of the table token.
-        - 'span_num': (Optional) The span number of the table token.
-        - 'line_num': (Optional) The line number of the table token.
-        - 'block_num': (Optional) The block number of the table token.
-    """
-    # TODO(yuming): update table_tokens from List[Dict] to List[TABLE_TOKEN]
-    # where TABLE_TOKEN will be a data class defined in unstructured-inference
+    ocr_layout = get_ocr_layout_from_image(
+        image,
+        ocr_languages=ocr_languages,
+        ocr_agent=ocr_agent,
+    )
     table_tokens = []
     for ocr_region in ocr_layout:
-        if ocr_region.bbox.is_in(
-            table_element.bbox,
-            error_margin=env_config.TABLE_TOKEN_ERROR_MARGIN,
-        ):
-            table_tokens.append(
-                {
-                    "bbox": [
-                        # token bound box is relative to table element
-                        ocr_region.bbox.x1 - table_element.bbox.x1,
-                        ocr_region.bbox.y1 - table_element.bbox.y1,
-                        ocr_region.bbox.x2 - table_element.bbox.x1,
-                        ocr_region.bbox.y2 - table_element.bbox.y1,
-                    ],
-                    "text": ocr_region.text,
-                },
-            )
+        table_tokens.append(
+            {
+                "bbox": [
+                    ocr_region.bbox.x1,
+                    ocr_region.bbox.y1,
+                    ocr_region.bbox.x2,
+                    ocr_region.bbox.y2,
+                ],
+                "text": ocr_region.text,
+            }
+        )
 
     # 'table_tokens' is a list of tokens
     # Need to be in a relative reading order
@@ -496,7 +470,7 @@ def get_ocr_layout_tesseract(
         text_height < env_config.TESSERACT_MIN_TEXT_HEIGHT
         or text_height > env_config.TESSERACT_MAX_TEXT_HEIGHT
     ):
-        # rounding avoids unnecessary precision and potential numerical issues assocaited
+        # rounding avoids unnecessary precision and potential numerical issues associated
         # with numbers very close to 1 inside cv2 image processing
         zoom = np.round(env_config.TESSERACT_OPTIMUM_TEXT_HEIGHT / text_height, 1)
         ocr_df = unstructured_pytesseract.image_to_data(