+++ unstructured/partition/pdf.py
@@ -282,6 +282,7 @@ def partition_pdf_or_image(
             file=file,
             is_image=is_image,
             infer_table_structure=infer_table_structure,
+            extract_images_in_pdf=extract_images_in_pdf,
         )
         != "ocr_only"
     ):
@@ -305,6 +306,7 @@ def partition_pdf_or_image(
         is_image=is_image,
         infer_table_structure=infer_table_structure,
         pdf_text_extractable=pdf_text_extractable,
+        extract_images_in_pdf=extract_images_in_pdf,
     )
 
     if strategy == "hi_res":